requirements.txt
Flask==3.1.0
playwright==1.52.0
PyNaCl==1.5.0
pytest==8.3.5
